requirements.txt
google-genai>=1.29.0
faiss-cpu>=1.7.4
python-dotenv==1.0.0
numpy>=1.24.3
pydantic>=2.0.0
pytest>=7.0.0
pytest-cov>=4.0.0
PyMuPDF>=1.26.0